Check for size before renaming .saving temp file.
git-svn-id: http://svn.vuze.com/public/client@3028 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/torrent/impl/TOTorrentImpl.java
@@ -130,10 +130,12 @@ TOTorrentImpl
               bos = null;
               
               //only use newly saved file if it got this far, i.e. it was written successfully
-              if ( output_file.exists() ) {
-                output_file.delete();
+              if ( temp.length() > 1L ) {
+                if ( output_file.exists() ) {
+                	output_file.delete();
+                }
+                temp.renameTo( output_file );
               }
-              temp.renameTo( output_file );
               
               return null;
